requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
pydantic
dotenv
pydantic[email]
